+++ plugin/src/main/kotlin/de/fayard/BuildSrcVersionsTask.kt
@@ -68,7 +68,7 @@ open class BuildSrcVersionsTask : DefaultTask() {
             KOTLIN_OBJECT -> false
             else -> return
         }
-        val versions = unsortedParsedDependencies.sortedBeautifullyBy { it.versionName }
+        val versions = unsortedParsedDependencies.sortedBeautifullyBy(extension.orderBy) { it.versionName }
 
         val kotlinPoetry: KotlinPoetry = kotlinpoet(versions, dependencyGraph.gradle, extension, computeIndent())
 
@@ -109,7 +109,7 @@ open class BuildSrcVersionsTask : DefaultTask() {
         }
 
         val dependencies = (unsortedParsedDependencies + specialDependencies)
-            .sortedBeautifullyBy { it.versionProperty }
+            .sortedBeautifullyBy(extension.orderBy) { it.versionProperty }
             .distinctBy { it.versionProperty }
 
         if (versionsOnlyMode == GRADLE_PROPERTIES) {
@@ -155,7 +155,7 @@ open class BuildSrcVersionsTask : DefaultTask() {
         val projectExtension = project.extensions.getByType<BuildSrcVersionsExtension>() as BuildSrcVersionsExtensionImpl
         this._extension = projectExtension.defensiveCopy()
         action.execute(this._extension)
-        PluginConfig.useRefreshVersions = project.hasProperty("plugin.de.fayard.buildSrcVersions")
+        PluginConfig.useRefreshVersions = project.hasProperty("plugin.de.fayard.buildSrcVersions") || project.hasProperty("plugin.de.refreshVersions")
     }
 
     private fun computeIndent(): String {