+++ plugin/src/main/kotlin/de/fayard/internal/KotlinPoetry.kt
@@ -8,6 +8,8 @@ import com.squareup.kotlinpoet.PropertySpec
 import com.squareup.kotlinpoet.TypeSpec
 import com.squareup.kotlinpoet.asClassName
 import de.fayard.BuildSrcVersionsExtension
+import de.fayard.OrderBy
+import de.fayard.OrderBy.*
 import org.gradle.plugin.use.PluginDependenciesSpec
 import org.gradle.plugin.use.PluginDependencySpec
 
@@ -62,10 +64,14 @@ fun kotlinpoet(
 }
 
 // https://github.com/jmfayard/buildSrcVersions/issues/65
-fun <T> List<T>.sortedBeautifullyBy(exceptIf: Boolean = false, selection: (T) -> String?) : List<T> =
-    this.filterNot { selection(it) == null }
+fun List<Dependency>.sortedBeautifullyBy(orderBy: OrderBy, selection: (Dependency) -> String?) : List<Dependency> {
+    val unsorted = this.filterNot { selection(it) == null }
         .sortedBy { selection(it)!! }
-        .sortedByDescending { selection(it)!!.length }
+    return when(orderBy) {
+        GROUP_AND_LENGTH -> unsorted.sortedByDescending { selection(it)!!.length }.sortedBy { it.mode }
+        GROUP_AND_ALPHABETICAL -> unsorted.sortedBy { it.mode }
+    }
+}
 
 fun FileSpec.Builder.addMaybeBuildSrcVersions(versions: List<Dependency>, extension: BuildSrcVersionsExtension) {
     versions.firstOrNull {
@@ -151,7 +157,7 @@ fun List<Dependency>.checkModeAndNames(useFdqnByDefault: List<String>): List<Dep
         d.escapedName = PluginConfig.escapeVersionsKt(
             when (d.mode) {
                 VersionMode.MODULE -> d.name
-                VersionMode.GROUP -> d.group
+                VersionMode.GROUP -> d.groupOrVirtualGroup()
                 VersionMode.GROUP_MODULE -> "${d.group}_${d.name}"
             }
         )
@@ -166,10 +172,11 @@ fun List<Dependency>.orderDependencies(): List<Dependency> {
 
 
 fun List<Dependency>.findCommonVersions(): List<Dependency> {
-    val map = groupBy { d -> d.group }
+    val map = groupBy { d: Dependency -> d.groupOrVirtualGroup() }
     for (deps in map.values) {
-        val groupTogether = deps.size > 1  && deps.map { it.version }.distinct().size == 1
-        if (groupTogether) {
+        val sameVersions = deps.map { it.version }.distinct().size == 1
+        val hasVirtualGroup = deps.any { it.groupOrVirtualGroup() != it.group }
+        if (sameVersions && (hasVirtualGroup || deps.size > 1)) {
             deps.forEach { d -> d.mode = VersionMode.GROUP }
         }
     }