+++ plugin/src/main/kotlin/de/fayard/internal/PluginConfig.kt
@@ -15,9 +15,8 @@ import java.io.File
 object PluginConfig {
 
 
-
-    const val PLUGIN_ID = "de.fayard.buildSrcVersions"
-    const val PLUGIN_VERSION = "0.7.0" // plugin.de.fayard.buildSrcVersions
+    const val PLUGIN_ID = "de.fayard.refreshVersions"
+    const val PLUGIN_VERSION = "0.7.0" // plugin.de.fayard.refreshVersions
     const val GRADLE_VERSIONS_PLUGIN_ID = "com.github.ben-manes.versions"
     const val GRADLE_VERSIONS_PLUGIN_VERSION = "0.25.0" // Sync with plugin/build.gradle.kts
     const val EXTENSION_NAME = "buildSrcVersions"
@@ -48,23 +47,36 @@ object PluginConfig {
      * Gradle properties can be set either in "gradle.properties" or from the command-line with
      *      $ ./gradlew -Pversion.kotlin.stdlib=1.3.50
      *  **/
-    fun considerGradleProperties(group: String, module: String): List<String> = listOf(
+    fun considerGradleProperties(group: String, module: String): List<String> = listOfNotNull(
         "version.$group..$module",
+        Dependency.virtualGroup(Dependency(group = group, name = module)),
         "version.$group",
         "version.$module"
     )
 
+    /**
+     * We want to treat all "org.getbrains.kotlinx:kotlinx-coroutines-*" as if they were a maven group
+     * with one common version, but different from org.jetbrains.kotlinx:kotlinx-serialization*
+     * For now this list is not part of the public API but feel free to add feedback that you need it.
+     * Add your use case here https://github.com/jmfayard/buildSrcVersions/issues/102
+     ***/
+    val virtualGroups : MutableList<String> = mutableListOf(
+        "org.jetbrains.kotlinx.kotlinx-coroutines",
+        "org.jetbrains.kotlinx.kotlinx-serialization"
+    )
+
+
     @JvmStatic
     fun versionPropertyFor(d: Dependency): String = when (d.mode) {
         MODULE -> d.name
-        GROUP -> d.group
+        GROUP -> d.groupOrVirtualGroup()
         GROUP_MODULE -> "${d.group}..${d.name}"
     }
 
     fun versionKtFor(d: Dependency): String = escapeVersionsKt(
         when (d.mode) {
             MODULE -> d.name
-            GROUP -> d.group
+            GROUP -> d.groupOrVirtualGroup()
             GROUP_MODULE -> "${d.group}:${d.name}"
         }
     )