+++ sample-groovy/buildSrc/src/main/kotlin/Libs.kt
@@ -7,14 +7,21 @@ import kotlin.String
  *   `$ ./gradlew buildSrcVersions`
  */
 object Libs {
+    const val androidx_annotation_annotation: String = "androidx.annotation:annotation:" +
+        Versions.androidx_annotation_annotation
+
+    const val org_jetbrains_annotation: String = "org.jetbrains:annotation:" +
+        Versions.org_jetbrains_annotation
+
+    /**
+     * http://code.google.com/p/google-guice/
+     */
+    const val com_google_inject_guice: String = "com.google.inject:guice:" +
+        Versions.com_google_inject_guice
+
     const val com_gradle_build_scan_gradle_plugin: String =
-            "com.gradle.build-scan:com.gradle.build-scan.gradle.plugin:2.4.2"
+        "com.gradle.build-scan:com.gradle.build-scan.gradle.plugin:" +
+            Versions.com_gradle_build_scan_gradle_plugin
 
-    const val androidx_annotation_annotation: String = "androidx.annotation:annotation:1.1.0"
-
-    const val org_jetbrains_annotation: String = "org.jetbrains:annotation:17.0.0"
-
-    const val com_google_inject_guice: String = "com.google.inject:guice:2.0"
-
-    const val guava: String = "com.google.guava:guava:15.0"
+    const val guava: String = "com.google.guava:guava:" + Versions.guava
 }